LayoutTests/accessibility/gtk/text-for-range-formatted-expected.txt

This is a test.

Hello world.
This sentence is green.
This one is not.
This tests the ability to get element text for a specified range.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS paragraph.stringForRange(0, -1) is 'This is a test.'
PASS block.stringForRange(0, 53) is 'Hello world.\nThis sentence is green.\nThis one is not.'
PASS successfullyParsed is true

TEST COMPLETE
